Migrate android device repo to light service
ProjectComponents are deprecated, migrating to services.
