+++ android/tooling/rib-intellij-plugin/src/main/kotlin/com/uber/intellij/plugin/android/rib/ui/RibHierarchyRootNodeDescriptor.kt
@@ -16,10 +16,11 @@
 package com.uber.intellij.plugin.android.rib.ui
 
 import com.intellij.icons.AllIcons
+import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.roots.ui.util.CompositeAppearance
 import com.intellij.psi.PsiElement
-import com.uber.intellij.plugin.android.rib.AndroidDeviceRepositoryComponent
+import com.uber.intellij.plugin.android.rib.AndroidDeviceRepository
 import com.uber.intellij.plugin.android.rib.RibHierarchyBrowser
 import com.uber.intellij.plugin.android.rib.RibProjectComponent
 import com.uber.intellij.plugin.android.rib.io.RibHost
@@ -32,6 +33,8 @@ class RibHierarchyRootNodeDescriptor(
   val ribHost: RibHost,
   private val status: RibHierarchyBrowser.Status
 ) : RibHierarchyDescriptor(project, null, element, true) {
+
+  private val deviceRepository = project.service<AndroidDeviceRepository>()
 
   companion object {
     /** Label used when android bridge is not connected */
@@ -50,7 +53,7 @@ class RibHierarchyRootNodeDescriptor(
   }
 
   override fun updateText(text: CompositeAppearance) {
-    if (!AndroidDeviceRepositoryComponent.getInstance(project).isBridgeConnected()) {
+    if (!deviceRepository.isBridgeConnected()) {
       text.ending.addText(LABEL_NO_BRIDGE)
       return
     }
